+++ src/godot_pool_string_array.lua
@@ -88,6 +88,9 @@ local methods = {
 	varianttype = VariantType.PoolStringArray,
 
 	--- Get the string at `index`.
+	-- Unlike Lua tables, indices start at 0 instead of 1.
+	-- For 1-based indexing, use the idiom `array[index]` instead.
+	--
 	-- If `index` is invalid (`index < 0` or `index >= size()`), the application will crash.
 	-- For a safe version that returns `nil` if `index` is invalid, use `safe_get` or the idiom `array[index]` instead.
 	-- @function get
@@ -98,14 +101,20 @@ local methods = {
 		return ffi_gc(api.godot_pool_string_array_get(self, index), api.godot_string_destroy)
 	end,
 	--- Get the string at `index`.
+	-- Unlike Lua tables, indices start at 0 instead of 1.
+	-- For 1-based indexing, use the idiom `array[index]` instead.
+	--
 	-- The idiom `array[index]` also calls this method.
 	-- @function safe_get
 	-- @tparam int index
 	-- @treturn[1] String
 	-- @treturn[2] nil  If index is invalid (`index < 0` or `index >= size()`)
 	-- @see get
-	safe_get = Array.safe_get,
+	safe_get = array_safe_get,
 	--- Set a new string for `index`.
+	-- Unlike Lua tables, indices start at 0 instead of 1.
+	-- For 1-based indexing, use the idiom `array[index] = value` instead.
+	--
 	-- If `index` is invalid (`index < 0` or `index >= size()`), the application will crash.
 	-- For a safe approach that `resize`s if `index >= size()`, use `safe_set` or the idiom `array[index] = value` instead.
 	-- @function set
@@ -116,6 +125,9 @@ local methods = {
 		api.godot_pool_string_array_set(self, index, str(value))
 	end,
 	--- Set a new string for `index`.
+	-- Unlike Lua tables, indices start at 0 instead of 1.
+	-- For 1-based indexing, use the idiom `array[index] = value` instead.
+	--
 	-- If `index >= size()`, the array is `resize`d first.
 	-- The idiom `array[index] = value` also calls this method.
 	-- @function safe_set
@@ -123,7 +135,7 @@ local methods = {
 	-- @param value  New value, stringified with `GD.str`
 	-- @raise If `index < 0`
 	-- @see set
-	safe_set = Array.safe_set,
+	safe_set = array_safe_set,
 	--- Inserts a new element at a given position in the array.
 	-- The position must be valid, or at the end of the array (`index == size()`).
 	-- @function insert
@@ -161,9 +173,7 @@ local methods = {
 	--- Returns `true` if the array is empty.
 	-- @function empty
 	-- @treturn bool
-	empty = function(self)
-		return #self == 0
-	end,
+	empty = array_empty,
 	--- Returns the [Read](#Class_PoolStringArray_Read) access for the array.
 	-- @function read
 	-- @treturn Read
@@ -201,7 +211,7 @@ end
 -- @function join
 -- @param[opt=""] delimiter  
 -- @treturn String
-methods.join = Array.join
+methods.join = array_join
 
 --- Static Functions.
 -- These don't receive `self` and should be called directly as `PoolStringArray.static_function(...)`
@@ -235,26 +245,30 @@ PoolStringArray = ffi_metatype('godot_pool_string_array', {
 	-- @param ...  Initial elements, added with `push_back`
 	-- @treturn PoolStringArray
 	__new = function(mt, ...)
-		local self = ffi.new(mt)
+		local self = ffi_new(mt)
 		api.godot_pool_string_array_new(self)
 		methods.push_back(self, ...)
 		return self
 	end,
 	__gc = api.godot_pool_string_array_destroy,
-	--- Returns method named `index` or the result of `safe_get`.
+	--- Returns method named `index` or the result of `safe_get(index - 1)`.
+	-- 
+	-- Like Lua tables, indices start at 1. For 0-based indexing, call `get` or
+	-- `safe_get` directly.
 	-- @function __index
 	-- @param index
 	-- @return Method or element or `nil`
 	-- @see safe_get
-	__index = function(self, index)
-		return methods[index] or methods.safe_get(self, index)
-	end,
-	--- Alias for `safe_set`.
+	__index = array_generate__index(methods),
+	--- Alias for `safe_set(index - 1, value)`.
+	--
+	-- Like Lua tables, indices start at 1. For 0-based indexing, call `set` or
+	-- `safe_set` directly.
 	-- @function __newindex
 	-- @tparam int index
 	-- @param value
 	-- @see safe_set
-	__newindex = methods.safe_set,
+	__newindex = array__newindex,
 	--- Returns a Lua string representation of this array.
 	-- @function __tostring
 	-- @treturn string
@@ -269,9 +283,7 @@ PoolStringArray = ffi_metatype('godot_pool_string_array', {
 	-- @function __len
 	-- @treturn int
 	-- @see size
-	__len = function(self)
-		return methods.size(self)
-	end,
+	__len = array__len,
 	--- Returns an iterator for array's elements, called by the idiom `ipairs(array)`.
 	-- @usage
 	--     for i, str in ipairs(array) do
@@ -280,10 +292,13 @@ PoolStringArray = ffi_metatype('godot_pool_string_array', {
 	-- @function __ipairs
 	-- @treturn function
 	-- @treturn PoolStringArray  self
+	-- @treturn int  0
 	__ipairs = array_ipairs,
 	--- Alias for `__ipairs`, called by the idiom `pairs(array)`.
 	-- @function __pairs
 	-- @treturn function
 	-- @treturn PoolStringArray  self
+	-- @treturn int  0
+	-- @see __ipairs
 	__pairs = array_ipairs,
 })