+++ src/test/extras/parse_error.lua
@@ -1,0 +1,1 @@
+hello, this is not a Lua script =O