+++ src/test/extras/valid_script.lua
@@ -1,0 +1,5 @@
+local Script = {
+	extends = 'Node'
+}
+
+return Script