+++ src/test/extras/valid_script_class_wrapper.lua
@@ -1,0 +1,6 @@
+local Script = {
+	extends = Node
+}
+
+return Script
+